requirements/base.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.39.1
cron-descriptor==1.4.5
cryptography==44.0.3
defusedxml==0.7.1
Django==4.2.15
django-appconf==1.1.0
django-celery-beat==2.6.0
django-celery-email==3.0.0
django-countries==7.6.1
django-filter==24.3
django-phonenumber-field==8.0.0
django-templated-mail==1.1.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.2.3
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
oauthlib==3.2.2
phonenumbers==8.13.42
pillow==10.4.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
loguru==0.7.2
redis==5.0.3
celery==5.3.6
flower==2.0.1
django-redis==5.4.0